speedup tests
