+++ sqs/pubsub_test.go
@@ -21,6 +21,8 @@ import (
 )
 
 func TestPubSub(t *testing.T) {
+	t.Parallel()
+
 	tests.TestPubSub(
 		t,
 		tests.Features{
@@ -34,21 +36,9 @@ func TestPubSub(t *testing.T) {
 	)
 }
 
-func TestPubSub_stress(t *testing.T) {
-	tests.TestPubSubStressTest(
-		t,
-		tests.Features{
-			ConsumerGroups:      false,
-			ExactlyOnceDelivery: false,
-			GuaranteedOrder:     true,
-			Persistent:          true,
-		},
-		createPubSub,
-		createPubSubWithConsumerGroup,
-	)
-}
-
 func TestPublishSubscribe_with_GenerateQueueUrlResolver(t *testing.T) {
+	t.Parallel()
+
 	tests.TestPublishSubscribe(
 		t,
 		tests.TestContext{
@@ -112,6 +102,8 @@ func TestPublishSubscribe_with_GenerateQueueUrlResolver(t *testing.T) {
 }
 
 func TestPublishSubscribe_with_TransparentUrlResolver(t *testing.T) {
+	t.Parallel()
+
 	tests.TestPublishSubscribe(
 		t,
 		tests.TestContext{
@@ -175,6 +167,8 @@ func TestPublishSubscribe_with_TransparentUrlResolver(t *testing.T) {
 }
 
 func TestPublishSubscribe_batching(t *testing.T) {
+	t.Parallel()
+
 	tests.TestPublishSubscribe(
 		t,
 		tests.TestContext{
@@ -231,6 +225,8 @@ func TestPublishSubscribe_batching(t *testing.T) {
 }
 
 func TestPublishSubscribe_creating_queue_with_different_settings_should_be_idempotent(t *testing.T) {
+	t.Parallel()
+
 	logger := watermill.NewStdLogger(false, false)
 
 	sub1, err := sqs.NewSubscriber(sqs.SubscriberConfig{
@@ -264,6 +260,8 @@ func TestPublishSubscribe_creating_queue_with_different_settings_should_be_idemp
 }
 
 func TestPublisher_GetOrCreateQueueUrl_is_idempotent(t *testing.T) {
+	t.Parallel()
+
 	pub, _ := createPubSub(t)
 
 	topicName := watermill.NewUUID()
@@ -280,6 +278,8 @@ func TestPublisher_GetOrCreateQueueUrl_is_idempotent(t *testing.T) {
 }
 
 func TestSubscriber_doesnt_hang_when_queue_doesnt_exist(t *testing.T) {
+	t.Parallel()
+
 	cfg := newAwsConfig(t)
 
 	_, sub := createPubSubWithConfig(
@@ -315,6 +315,8 @@ func TestSubscriber_doesnt_hang_when_queue_doesnt_exist(t *testing.T) {
 }
 
 func TestPublisher_do_not_create_queue(t *testing.T) {
+	t.Parallel()
+
 	cfg := newAwsConfig(t)
 
 	pub, _ := createPubSubWithConfig(
